rename: TapPanel -> TOCHeading 이름변경
